requirements.txt
python-dotenv
pyodbc
realtime
retry
simplejson
supabase
websockets